backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "total-llm"
version = "2.0.0"
description = "Total-LLM v2 backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.124",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.10",
  "pydantic-settings>=2.10",
  "langgraph>=1.0",
  "langgraph-checkpoint-postgres",
  "langchain>=1.0",
  "langchain-openai>=1.0",
  "langchain-community",
  "openai>=1.0",
  "sentence-transformers>=5.0",
  "qdrant-client>=1.16",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.3",
  "redis>=7.0",
  "aiohttp>=3.9",
  "python-multipart",
  "websockets>=12.0",
  "python-jose[cryptography]",
  "passlib",
  "bcrypt==4.0.1",
  "pypdf>=3.17",
  "python-docx>=1.1",
  "markdown>=3.5",
  "rank-bm25>=0.2",
  "reportlab>=4.0",
  "pyyaml>=6.0",
  "python-dotenv",
  "fastmcp>=3.0",
  "beautifulsoup4>=4.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src/total_llm"]
